test: move fixtures
