+++ src/record/export/printers/printAsCsv/__tests__/fixtures/index/common/fields.ts
@@ -1,4 +1,4 @@
-import { FieldsJson } from "../../../../../../../kintone/types";
+import { FieldsJson } from "../../../../../../../../kintone/types";
 
 export const fieldsJson: FieldsJson = {
   revision: "29",
@@ -38,81 +38,6 @@ export const fieldsJson: FieldsJson = {
       code: "creator",
       label: "creator",
       noLabel: false,
-    },
-    subTable: {
-      type: "SUBTABLE",
-      code: "subTable",
-      noLabel: false,
-      label: "subTable",
-      fields: {
-        subTableText: {
-          type: "SINGLE_LINE_TEXT",
-          code: "subTableText",
-          label: "subTableText",
-          noLabel: false,
-          required: false,
-          minLength: "",
-          maxLength: "",
-          expression: "",
-          hideExpression: false,
-          unique: false,
-          defaultValue: "",
-        },
-        subTableCheckbox: {
-          type: "CHECK_BOX",
-          code: "subTableCheckbox",
-          label: "subTableCheckbox",
-          noLabel: false,
-          required: false,
-          options: {
-            st_sample2: {
-              label: "st_sample2",
-              index: "1",
-            },
-            st_sample1: {
-              label: "st_sample1",
-              index: "0",
-            },
-          },
-          defaultValue: [],
-          align: "HORIZONTAL",
-        },
-        subTableFile: {
-          type: "FILE",
-          code: "subTableFile",
-          label: "subTableFile",
-          noLabel: false,
-          required: false,
-          thumbnailSize: "150",
-        },
-        userSelect: {
-          type: "USER_SELECT",
-          code: "userSelect",
-          label: "userSelect",
-          noLabel: false,
-          required: false,
-          entities: [],
-          defaultValue: [],
-        },
-        organizationSelect: {
-          type: "ORGANIZATION_SELECT",
-          code: "organizationSelect",
-          label: "organizationSelect",
-          noLabel: false,
-          required: false,
-          entities: [],
-          defaultValue: [],
-        },
-        groupSelect: {
-          type: "GROUP_SELECT",
-          code: "groupSelect",
-          label: "groupSelect",
-          noLabel: false,
-          required: false,
-          entities: [],
-          defaultValue: [],
-        },
-      },
     },
     Assignee: {
       type: "STATUS_ASSIGNEE",
@@ -268,6 +193,62 @@ export const fieldsJson: FieldsJson = {
         },
       },
       defaultValue: [],
+    },
+    userSelect: {
+      type: "USER_SELECT",
+      code: "userSelect",
+      label: "userSelect",
+      noLabel: false,
+      required: false,
+      entities: [],
+      defaultValue: [],
+    },
+    organizationSelect: {
+      type: "ORGANIZATION_SELECT",
+      code: "organizationSelect",
+      label: "organizationSelect",
+      noLabel: false,
+      required: false,
+      entities: [],
+      defaultValue: [],
+    },
+    groupSelect: {
+      type: "GROUP_SELECT",
+      code: "groupSelect",
+      label: "groupSelect",
+      noLabel: false,
+      required: false,
+      entities: [],
+      defaultValue: [],
+    },
+    date: {
+      type: "DATE",
+      code: "date",
+      label: "date",
+      noLabel: false,
+      required: false,
+      unique: false,
+      defaultValue: "",
+      defaultNowValue: true,
+    },
+    dateTime: {
+      type: "DATETIME",
+      code: "dateTime",
+      label: "dateTime",
+      noLabel: false,
+      required: false,
+      unique: false,
+      defaultValue: "",
+      defaultNowValue: true,
+    },
+    time: {
+      type: "TIME",
+      code: "time",
+      label: "time",
+      noLabel: false,
+      required: false,
+      defaultValue: "",
+      defaultNowValue: true,
     },
   },
 };
