+++ src/record/export/printers/printAsCsv/__tests__/fixtures/index/common/index.ts
@@ -1,11 +1,10 @@
-import { TestPattern } from "../../index.test";
+import { TestPattern } from "../../../index.test";
 import { expectedCsv } from "./expected";
 import { input } from "./input";
 import { fieldsJson } from "./fields";
 
 export const pattern: TestPattern = {
-  description:
-    "should convert kintone records to csv string correctly when SUBTABLE included without attachmentsDir option",
+  description: "should convert kintone records to csv string correctly",
   fieldsJson: fieldsJson,
   input: input,
   attachmentsDir: "attachmentsDir",
