+++ src/record/export/printers/printAsCsv/__tests__/fixtures/index/withEmptySubtable/expected.ts
