+++ src/record/export/printers/printAsCsv/__tests__/fixtures/index/withEmptySubtable/index.ts
@@ -1,10 +1,11 @@
-import { TestPattern } from "../../index.test";
+import { TestPattern } from "../../../index.test";
 import { expectedCsv } from "./expected";
 import { input } from "./input";
 import { fieldsJson } from "./fields";
 
 export const pattern: TestPattern = {
-  description: "should convert kintone records to csv string correctly",
+  description:
+    "should convert kintone records to csv string correctly when SUBTABLE value is an empty array",
   fieldsJson: fieldsJson,
   input: input,
   attachmentsDir: "attachmentsDir",
