+++ src/record/export/printers/printAsCsv/__tests__/fixtures/index/withFileAndAttachmentsDir/input.ts
@@ -1,4 +1,4 @@
-import { KintoneRecord } from "../../../../../types/record";
+import { KintoneRecord } from "../../../../../../types/record";
 
 export const input: KintoneRecord[] = [
   {
@@ -73,9 +73,24 @@ export const input: KintoneRecord[] = [
       type: "__ID__",
       value: "9",
     },
-    subTable: {
-      type: "SUBTABLE",
-      value: [],
+    file: {
+      type: "FILE",
+      value: [
+        {
+          contentType: "text/plain",
+          fileKey: "test-file-key",
+          name: "test.txt",
+          size: "12345",
+          localFilePath: "file-9/test.txt",
+        },
+        {
+          contentType: "text/plain",
+          fileKey: "test-file-key",
+          name: "test.txt",
+          size: "12345",
+          localFilePath: "file-9/test (1).txt",
+        },
+      ],
     },
   },
 ];
