+++ src/record/export/printers/printAsCsv/__tests__/fixtures/index/withFileAndNoAttachmentsDir/expected.ts
