+++ src/record/export/printers/printAsCsv/__tests__/fixtures/index/withFileAndNoAttachmentsDir/input.ts
@@ -1,4 +1,4 @@
-import { KintoneRecord } from "../../../../../types/record";
+import { KintoneRecord } from "../../../../../../types/record";
 
 export const input: KintoneRecord[] = [
   {
@@ -79,16 +79,14 @@ export const input: KintoneRecord[] = [
         {
           contentType: "text/plain",
           fileKey: "test-file-key",
-          name: "test.txt",
           size: "12345",
-          localFilePath: "file-9/test.txt",
+          name: "test.txt",
         },
         {
           contentType: "text/plain",
           fileKey: "test-file-key",
-          name: "test.txt",
           size: "12345",
-          localFilePath: "file-9/test (1).txt",
+          name: "test.txt",
         },
       ],
     },
