+++ src/record/export/printers/printAsCsv/__tests__/fixtures/index/withSubtableAndFileAndAttachmentsDir/index.ts
@@ -1,11 +1,11 @@
-import { TestPattern } from "../../index.test";
+import { TestPattern } from "../../../index.test";
 import { expectedCsv } from "./expected";
 import { input } from "./input";
 import { fieldsJson } from "./fields";
 
 export const pattern: TestPattern = {
   description:
-    "should convert kintone records to csv string correctly when SUBTABLE value is an empty array",
+    "should convert kintone records to csv string correctly when SUBTABLE included with attachmentsDir option",
   fieldsJson: fieldsJson,
   input: input,
   attachmentsDir: "attachmentsDir",
