+++ src/record/export/printers/printAsCsv/__tests__/fixtures/index/withSubtableAndFileAndNoAttachmentsDir/fields.ts
@@ -1,4 +1,4 @@
-import { FieldsJson } from "../../../../../../../kintone/types";
+import { FieldsJson } from "../../../../../../../../kintone/types";
 
 export const fieldsJson: FieldsJson = {
   revision: "29",
