+++ src/record/export/printers/printAsCsv/__tests__/fixtures/index/withSubtableAndFileAndNoAttachmentsDir/index.ts
@@ -1,11 +1,11 @@
-import { TestPattern } from "../../index.test";
+import { TestPattern } from "../../../index.test";
 import { expectedCsv } from "./expected";
 import { input } from "./input";
 import { fieldsJson } from "./fields";
 
 export const pattern: TestPattern = {
   description:
-    "should convert kintone records to csv string correctly when FILE included and with attachmentsDir option",
+    "should convert kintone records to csv string correctly when SUBTABLE included without attachmentsDir option",
   fieldsJson: fieldsJson,
   input: input,
   attachmentsDir: "attachmentsDir",
